requirements.txt
asgiref==3.5.2
coverage==6.4
Django==4.1.2
django-ratelimit==3.0.1
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
